+++ src/Services/Auth/SimpleJwtGuard.php
@@ -7,8 +7,15 @@ use Illuminate\Contracts\Auth\Guard;
 use Illuminate\Support\Traits\Macroable;
 use Illuminate\Contracts\Auth\UserProvider;
 use Illuminate\Http\Request;
-use aliirfaan\LaravelSimpleJwt\Services\JwtHelperService;
 use Illuminate\Contracts\Auth\Authenticatable as AuthenticatableContract;
+use Illuminate\Contracts\Events\Dispatcher;
+use Illuminate\Auth\Events\Attempting;
+use Illuminate\Auth\Events\Authenticated;
+use Illuminate\Auth\Events\Failed;
+use Illuminate\Auth\Events\Login;
+use Illuminate\Auth\Events\Validated;
+use aliirfaan\LaravelSimpleJwt\Services\JwtHelperService;
+use aliirfaan\LaravelSimpleJwt\Contracts\JwtServiceInterface;
 
 class SimpleJwtGuard implements Guard
 {
@@ -43,20 +50,45 @@ class SimpleJwtGuard implements Guard
     protected $request;
 
     /**
+     * The event dispatcher instance.
+     *
+     * @var \Illuminate\Contracts\Events\Dispatcher
+     */
+    protected $events;
+
+    /**
+     * Simple JWT profile to load
+     *
+     * @var string
+     */
+    protected $profile;
+
+    /**
      * Create a new authentication guard.
      *
      * @param  string  $name
      * @param  \Illuminate\Contracts\Auth\UserProvider  $provider
      * @param  \Illuminate\Contracts\Session\Session  $session
+     * @param  string  $profile
+     * @param  \Illuminate\Contracts\Events\Dispatcher|null  $dispatcher
      * @param  \Symfony\Component\HttpFoundation\Request|null  $request
      * @return void
      */
-    public function __construct($name, UserProvider $provider, Request $request = null)
+    public function __construct($name, UserProvider $provider, JwtServiceInterface $jwtService = null, $profile = 'default', Dispatcher $dispatcher = null, Request $request = null)
     {
         $this->name = $name;
-        $this->jwtService = new JwtHelperService();
+        $this->events = $dispatcher;
+
+        $this->jwtService = null;
+        if (is_null($jwtService)) {
+            $this->jwtService = new JwtHelperService();
+        } else {
+            $this->jwtService = $jwtService;
+        }
+
         $this->request = $request;
         $this->provider = $provider;
+        $this->profile = $profile;
     }
 
     /**
@@ -71,11 +103,10 @@ class SimpleJwtGuard implements Guard
         }
 
         $token = $this->request->bearerToken();
-        $verifyToken = $this->jwtService->verifyJwtToken($token);
+        $verifyToken = $this->jwtService->verifyJwtToken($token, $this->profile);
         if ($verifyToken['errors'] == null) {
             $tokenClaims = (array) $verifyToken['result'];
             $this->user = $this->provider->retrieveById($tokenClaims['sub']);
-
             return $this->user;
         }
 
@@ -105,10 +136,14 @@ class SimpleJwtGuard implements Guard
      */
     protected function hasValidCredentials($user, $credentials)
     {
-        $validated = ! is_null($user) && $this->provider->validateCredentials($user, $credentials);
+        try {
+            $validated = ! is_null($user) && $this->provider->validateCredentials($user, $credentials);
 
-        if ($validated) {
-            //$this->fireValidatedEvent($user);
+            if ($validated) {
+                $this->fireValidatedEvent($user);
+            }
+        } catch (\Exception $e) {
+            throw $e;
         }
 
         return $validated;
@@ -123,21 +158,27 @@ class SimpleJwtGuard implements Guard
      */
     public function attempt(array $credentials = [], $remember = false)
     {
-        //$this->fireAttemptEvent($credentials, $remember);
+        try {
+            $this->fireAttemptEvent($credentials, $remember);
 
-        $this->lastAttempted = $user = $this->provider->retrieveByCredentials($credentials);
+            $this->lastAttempted = $user = $this->provider->retrieveByCredentials($credentials);
 
-        // If an implementation of UserInterface was returned, we'll ask the provider
-        // to validate the user against the given credentials, and if they are in
-        // fact valid we'll log the users into the application and return true.
-        if ($this->hasValidCredentials($user, $credentials)) {
-            return $this->login($user, $remember);
+            // If an implementation of UserInterface was returned, we'll ask the provider
+            // to validate the user against the given credentials, and if they are in
+            // fact valid we'll log the users into the application and return true.
+            if ($this->hasValidCredentials($user, $credentials)) {
+                return $this->login($user, $remember);
+            }
+
+        } catch (\Exception $e) {
+            $this->fireFailedEvent($user, $credentials);
+            throw $e;
         }
 
         // If the authentication attempt fails we will fire an event so that the user
         // may be notified of any suspicious attempts to access their account from
         // an unrecognized user. A developer may listen to this event as needed.
-        //$this->fireFailedEvent($user, $credentials);
+        $this->fireFailedEvent($user, $credentials);
 
         return false;
     }
@@ -155,7 +196,7 @@ class SimpleJwtGuard implements Guard
         // If we have an event dispatcher instance set we will fire an event so that
         // any listeners will hook into the authentication events and run actions
         // based on the login and logout events fired from the guard instances.
-        //$this->fireLoginEvent($user, $remember);
+        $this->fireLoginEvent($user, $remember);
 
         // return token here
 
@@ -163,12 +204,157 @@ class SimpleJwtGuard implements Guard
         $tokenPayload = [
             'sub' => $user->getAuthIdentifier(),
         ];
-        $jwt = $this->jwtService->createJwtToken($tokenPayload);
+        $jwt = $this->jwtService->createJwtToken($tokenPayload, $this->profile);
+
+        $this->setUser($user);
+
+        $this->fireAuthenticatedEvent($user);
 
         return $jwt;
+    }
 
-        // refresh token
+    /**
+     * Get the event dispatcher instance.
+     *
+     * @return \Illuminate\Contracts\Events\Dispatcher
+     */
+    public function getDispatcher()
+    {
+        return $this->events;
+    }
 
-        //$this->setUser($user);
+    /**
+     * Set the event dispatcher instance.
+     *
+     * @param  \Illuminate\Contracts\Events\Dispatcher  $events
+     * @return void
+     */
+    public function setDispatcher(Dispatcher $events)
+    {
+        $this->events = $events;
+    }
+
+    /**
+     * Register an authentication attempt event listener.
+     *
+     * @param  mixed  $callback
+     * @return void
+     */
+    public function attempting($callback)
+    {
+        if (isset($this->events)) {
+            $this->events->listen(Events\Attempting::class, $callback);
+        }
+    }
+
+    /**
+     * Fire the attempt event with the arguments.
+     *
+     * @param  array  $credentials
+     * @param  bool  $remember
+     * @return void
+     */
+    protected function fireAttemptEvent(array $credentials, $remember = false)
+    {
+        if (isset($this->events)) {
+            $this->events->dispatch(new Attempting(
+                $this->name, $credentials, $remember
+            ));
+        }
+    }
+
+    /**
+     * Fires the validated event if the dispatcher is set.
+     *
+     * @param  \Illuminate\Contracts\Auth\Authenticatable  $user
+     * @return void
+     */
+    protected function fireValidatedEvent($user)
+    {
+        if (isset($this->events)) {
+            $this->events->dispatch(new Validated(
+                $this->name, $user
+            ));
+        }
+    }
+
+    /**
+     * Fire the login event if the dispatcher is set.
+     *
+     * @param  \Illuminate\Contracts\Auth\Authenticatable  $user
+     * @param  bool  $remember
+     * @return void
+     */
+    protected function fireLoginEvent($user, $remember = false)
+    {
+        if (isset($this->events)) {
+            $this->events->dispatch(new Login(
+                $this->name, $user, $remember
+            ));
+        }
+    }
+
+    /**
+     * Fire the authenticated event if the dispatcher is set.
+     *
+     * @param  \Illuminate\Contracts\Auth\Authenticatable  $user
+     * @return void
+     */
+    protected function fireAuthenticatedEvent($user)
+    {
+        if (isset($this->events)) {
+            $this->events->dispatch(new Authenticated(
+                $this->name, $user
+            ));
+        }
+    }
+
+    /**
+     * Fire the failed authentication attempt event with the given arguments.
+     *
+     * @param  \Illuminate\Contracts\Auth\Authenticatable|null  $user
+     * @param  array  $credentials
+     * @return void
+     */
+    protected function fireFailedEvent($user, array $credentials)
+    {
+        if (isset($this->events)) {
+            $this->events->dispatch(new Failed(
+                $this->name, $user, $credentials
+            ));
+        }
+    }
+
+    /**
+     * Get the current request instance.
+     *
+     * @return \Symfony\Component\HttpFoundation\Request
+     */
+    public function getRequest()
+    {
+        return $this->request ?: Request::createFromGlobals();
+    }
+
+    /**
+     * Set the current request instance.
+     *
+     * @param  \Symfony\Component\HttpFoundation\Request  $request
+     * @return $this
+     */
+    public function setRequest(Request $request)
+    {
+        $this->request = $request;
+
+        return $this;
+    }
+
+    /**
+     * Return the currently cached user.
+     *
+     * @return \Illuminate\Contracts\Auth\Authenticatable|null
+     */
+    public function getUser()
+    {
+        return $this->user;
     }
 }