+++ src/SimpleJwtServiceProvider.php
@@ -4,6 +4,7 @@ namespace aliirfaan\LaravelSimpleJwt;
 
 use Illuminate\Support\Facades\Auth;
 use aliirfaan\LaravelSimpleJwt\Services\Auth\SimpleJwtGuard;
+use aliirfaan\LaravelSimpleJwt\Services\JwtHelperService;
 
 class SimpleJwtServiceProvider extends \Illuminate\Support\ServiceProvider
 {
@@ -19,7 +20,7 @@ class SimpleJwtServiceProvider extends \Illuminate\Support\ServiceProvider
         });
 
         Auth::extend('simple-jwt-guard', function ($app, $name, array $config) {
-            return new SimpleJwtGuard($name, Auth::createUserProvider($config['provider']));
+            return new SimpleJwtGuard($name, Auth::createUserProvider($config['provider']), $config['jwt_class'], $config['profile'], $this->app['events'], $this->app['request']);
         });
     }